Fix build on armhf with gcc 7.
The abs caused a compiler error in that environment, because wchar_t is unsigned on armhf, subtracting signed from unsigned results in unsigned and std::abs does not define an unsigned variant.

Differential Revision: https://code.wildfiregames.com/D1122
Patch By: Peter Michael Green
Uploaded By: Ludovic



git-svn-id: https://svn.wildfiregames.com/public/ps/trunk@20869 3db68df2-c116-0410-a063-a993310a9797

+++ source/gui/CDropDown.cpp
@@ -352,7 +352,7 @@ InReaction CDropDown::ManuallyHandleEvent(const SDL_Event_* ev)
 					int diff = 0;
 					for (size_t j = 0; j < m_InputBuffer.length(); ++j)
 					{
-						diff = std::abs(pList->m_Items[i].GetRawString().LowerCase()[j] - (int)m_InputBuffer[j]);
+						diff = std::abs((int)(pList->m_Items[i].GetRawString().LowerCase()[j]) - (int)m_InputBuffer[j]);
 						if (diff == 0)
 							indexOfDifference = j+1;
 						else
